requirements.txt
beautifulsoup4
playwright
pillow
requests